Move RefillAmmo into components
